+++ NEWS.md
@@ -11,6 +11,8 @@ Changes:
 
 4. Updating RedShift driver to V2.1.0.9. Fixes error when uploading data.
 
+5. Adding `requiresTempEmulation()` and `assertTempEmulationSchemaSet()` function.
+
 
 Bugfixes: